refactor(core): move force layout helper to dom utils
