+++ packages/core/src/dom/utils/index.ts
@@ -1,5 +1,6 @@
 export type { StateAttrMap } from '../../core/ui/types';
 export { applyElementProps } from './element-props';
+export { forceLayout } from './layout';
 export { logMissingFeature } from './log';
 export { getPercentFromPointerEvent } from './pointer';
 export { applyStateDataAttrs, getStateDataAttrs } from './state-data-attrs';
